Explode the overloaded component
And now legacy goes for current, and not other way
around.
