--- impl/maven-core/src/main/java/org/apache/maven/internal/impl/DefaultPrompter.java
@@ -1,66 +1,0 @@
-/*
- * Licensed to the Apache Software Foundation (ASF) under one
- * or more contributor license agreements.  See the NOTICE file
- * distributed with this work for additional information
- * regarding copyright ownership.  The ASF licenses this file
- * to you under the Apache License, Version 2.0 (the
- * "License"); you may not use this file except in compliance
- * with the License.  You may obtain a copy of the License at
- *
- *   http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing,
- * software distributed under the License is distributed on an
- * "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY
- * KIND, either express or implied.  See the License for the
- * specific language governing permissions and limitations
- * under the License.
- */
-package org.apache.maven.internal.impl;
-
-import javax.inject.Inject;
-import javax.inject.Named;
-import javax.inject.Singleton;
-
-import java.util.List;
-
-import org.apache.maven.api.services.Prompter;
-import org.apache.maven.api.services.PrompterException;
-
-@Named
-@Singleton
-public class DefaultPrompter implements Prompter {
-    private final org.codehaus.plexus.components.interactivity.Prompter prompter;
-
-    @Inject
-    public DefaultPrompter(org.codehaus.plexus.components.interactivity.Prompter prompter) {
-        this.prompter = prompter;
-    }
-
-    @Override
-    public String prompt(String message, List<String> possibleValues, String defaultReply) throws PrompterException {
-        try {
-            return prompter.prompt(message, possibleValues, defaultReply);
-        } catch (Exception e) {
-            throw new PrompterException("Unable to call prompter", e);
-        }
-    }
-
-    @Override
-    public String promptForPassword(String message) throws PrompterException {
-        try {
-            return prompter.promptForPassword(message);
-        } catch (Exception e) {
-            throw new PrompterException("Unable to call prompter", e);
-        }
-    }
-
-    @Override
-    public void showMessage(String message) throws PrompterException {
-        try {
-            prompter.showMessage(message);
-        } catch (Exception e) {
-            throw new PrompterException("Unable to call prompter", e);
-        }
-    }
-}
